+++ packages/entities/entities-gateway-services/src/components/NewGatewayServiceForm.vue
@@ -453,6 +453,7 @@ import {
   demoStockService,
   demoWeatherService,
 } from '../constants'
+import useUrlValidators from 'src/composables/useUrlValidators'
 
 const emit = defineEmits<{
   (e: 'update', data: Record<string, any>): void,
@@ -509,6 +510,7 @@ const { i18nT, i18n: { t } } = composables.useI18n()
 const { getErrorFieldsFromError } = useErrors()
 const { axiosInstance } = useAxios(props.config?.axiosRequestConfig)
 const validators = useValidators()
+const { validateHost, validatePath, validatePort } = useUrlValidators()
 
 const fetchUrl = computed<string>(() => endpoints.form[props.config.app].edit)
 const formType = computed((): EntityBaseFormType => props.gatewayServiceId ? EntityBaseFormType.Edit : EntityBaseFormType.Create)
@@ -711,16 +713,6 @@ const handleTrySampleApi = (): void => {
   }
 }
 
-const validateHost = (host: string): string => {
-  // check if host is empty
-  if (!host || host.trim() === '') return 'host cannot be empty'
-
-  // check if a valid host (domain or ip address)
-  const hostnameRegex = /^(([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9-]*[a-zA-Z0-9])\.)*([A-Za-z0-9]|[A-Za-z0-9][A-Za-z0-9-]*[A-Za-z0-9])$|^\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}$/
-  if (!hostnameRegex.test(host)) return 'invalid host'
-  return ''
-}
-
 const validateProtocol = (protocol: string) => {
   // check if protocol is empty
   if (!protocol || protocol.trim() === '') return 'protocol cannot be empty'
@@ -734,51 +726,6 @@ const validateProtocol = (protocol: string) => {
 
   if (selectedProtocol === undefined) return 'protocol - value must be one of "http", "https", "grpc", "grpcs", "tcp", "udp", "tls", "tls_passthrough", "ws", "wss"'
 
-  return ''
-}
-
-const validatePort = (port: number | null | undefined | string): string => {
-  // Port is not required, so return empty string if not provided
-  if (port === null || port === undefined || port === '') {
-    return ''
-  }
-
-  // Convert to number if it's a string
-  const portNum = typeof port === 'string' ? parseInt(port, 10) : port
-
-  // Check if port is a valid number
-  if (isNaN(portNum)) {
-    return 'Port must be a number'
-  }
-
-  // Check if port is in valid range (0-65535)
-  if (portNum < 0 || portNum > 65535) {
-    return 'Port must be between 0 and 65535'
-  }
-
-  // Valid port
-  return ''
-}
-
-const validatePath = (path: string | null | undefined): string => {
-  // Path is not required, so return empty string if not provided
-  if (!path || path === '') {
-    return ''
-  }
-
-  // If path exists, it should start with '/'
-  if (!path.startsWith('/')) {
-    return 'Path must begin with /'
-  }
-
-  // This regex matches any character that is NOT allowed by RFC 3986
-  const invalidCharsRegex = /[^A-Za-z0-9\-._~:/?#[\]@!$&'()*+,;=%]/
-
-  if (invalidCharsRegex.test(path)) {
-    return 'Path should not include characters outside of the reserved list of RFC 3986'
-  }
-
-  // Valid path according to RFC 3986
   return ''
 }
