+++ packages/entities/entities-gateway-services/src/index.ts
@@ -1,5 +1,5 @@
 import GatewayServiceList from './components/GatewayServiceList.vue'
-import GatewayServiceForm from './components/GatewayServiceForm.vue'
+import GatewayServiceForm from './components/LegacyGatewayServiceForm.vue'
 import GatewayServiceConfigCard from './components/GatewayServiceConfigCard.vue'
 import NewGatewayServiceForm from './components/NewGatewayServiceForm.vue'
